+++ NetFabric.Hyperlinq/Aggregation/Count/Count.ReadOnlyList.cs
@@ -9,7 +9,7 @@ namespace NetFabric.Hyperlinq
             where TEnumerable : IReadOnlyList<TSource>
             => source.Count;
 
-        public static int Count<TEnumerable, TSource>(this TEnumerable source, Func<TSource, int, bool> predicate)
+        public static int Count<TEnumerable, TSource>(this TEnumerable source, Func<TSource, long, bool> predicate)
             where TEnumerable : IReadOnlyList<TSource>
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
