+++ NetFabric.Hyperlinq/Element/First/First.Array.cs
@@ -12,7 +12,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, int, bool> predicate) 
+        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, long, bool> predicate) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -26,7 +26,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index) 
+        public static ref TSource First<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -48,7 +48,7 @@ namespace NetFabric.Hyperlinq
             return ref source[0];
         }
 
-        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate) 
+        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -61,7 +61,7 @@ namespace NetFabric.Hyperlinq
             return ref Default<TSource>.Value;
         }
 
-        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index) 
+        public static ref readonly TSource FirstOrDefault<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index) 
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
 
@@ -85,11 +85,11 @@ namespace NetFabric.Hyperlinq
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static TSource? FirstOrNull<TSource>(this TSource[] source, Func<TSource, int, bool> predicate) 
+        public static TSource? FirstOrNull<TSource>(this TSource[] source, Func<TSource, long, bool> predicate) 
             where TSource : struct
             => FirstOrNull<TSource>(source, predicate, out var _); 
 
-        public static TSource? FirstOrNull<TSource>(this TSource[] source, Func<TSource, int, bool> predicate, out int index) 
+        public static TSource? FirstOrNull<TSource>(this TSource[] source, Func<TSource, long, bool> predicate, out int index) 
             where TSource : struct
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
